+++ drift_dev/lib/src/model/base_entity.dart
@@ -88,7 +88,7 @@ class ExistingRowClass {
             typeArguments: typeInstantiation,
             nullabilitySuffix: NullabilitySuffix.none,
           )
-          .getDisplayString(withNullability: options.nnbd);
+          .getDisplayString(withNullability: true);
     }
   }
 }