+++ drift_dev/lib/src/model/column.dart
@@ -161,7 +161,7 @@ class MoorColumn implements HasDeclaration, HasType {
     // We currently use nullable columns everywhere because it's not clear how
     // to express nullability in joins otherwise. When serializing data with
     // JsonTypeConverter it needs to match the nullability of the column
-    return options.nnbd && (!checkNullable || nullable) ? '$code?' : code;
+    return (!checkNullable || nullable) ? '$code?' : code;
   }
 
   SqlType sqlType() {