+++ drift_dev/lib/src/writer/database_writer.dart
@@ -70,7 +70,6 @@ class DatabaseWriter {
           getterName: entity.dbGetterName,
           returnType: tableClassName,
           code: '$tableClassName(this)',
-          options: scope.generationOptions,
         );
       } else if (entity is MoorTrigger) {
         writeMemoizedGetter(
@@ -79,7 +78,6 @@ class DatabaseWriter {
           returnType: 'Trigger',
           code: 'Trigger(${asDartLiteral(entity.createSql(scope.options))}, '
               '${asDartLiteral(entity.displayName)})',
-          options: scope.generationOptions,
         );
       } else if (entity is MoorIndex) {
         writeMemoizedGetter(
@@ -88,7 +86,6 @@ class DatabaseWriter {
           returnType: 'Index',
           code: 'Index(${asDartLiteral(entity.displayName)}, '
               '${asDartLiteral(entity.createSql(scope.options))})',
-          options: scope.generationOptions,
         );
       } else if (entity is MoorView) {
         writeMemoizedGetter(
@@ -96,7 +93,6 @@ class DatabaseWriter {
           getterName: entity.dbGetterName,
           returnType: entity.entityInfoName,
           code: '${entity.entityInfoName}(this)',
-          options: scope.generationOptions,
         );
       }
     }
@@ -112,7 +108,6 @@ class DatabaseWriter {
         getterName: getterName,
         returnType: typeName,
         code: '$typeName(this as $databaseImplName)',
-        options: scope.generationOptions,
       );
     }