+++ drift_dev/lib/src/writer/queries/result_set_writer.dart
@@ -44,7 +44,7 @@ class ResultSetWriter {
         final fieldName = nested.dartFieldName;
 
         if (nested.isNullable) {
-          typeName = scope.nullableType(typeName);
+          typeName += '?';
         }
 
         into.write('$modifier $typeName $fieldName;\n');
@@ -68,16 +68,14 @@ class ResultSetWriter {
 
     // write the constructor
     if (scope.options.rawResultSetData) {
-      into.write('$className({${scope.required} QueryRow row,');
+      into.write('$className({required QueryRow row,');
     } else {
       into.write('$className({');
     }
 
     for (final column in fieldNames) {
       if (nonNullableFields.contains(column)) {
-        into
-          ..write(scope.required)
-          ..write(' ');
+        into.write('required ');
       }
       into.write('this.$column,');
     }